requirements.txt
flask
rpi-lgpio